ident and fix

+++ src/Solido.java
@@ -1,10 +1,10 @@
 
 public class Solido {
-	protected float altura;
-	public float getAltura() {
+	protected double altura;
+	public double getAltura() {
 		return altura;
 	}
-	public void setAltura(float altura) {
+	public void setAltura(double altura) {
 		this.altura = altura;
 	}
